+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# REPLACE_GLOBALLY_WITH_NEXT_VERSION
+- PPI-199 - Improve webhook log messages
+
 # 2.0.1
 - PPI-171 - Message queue is now only used if there are iZettle Sales Channels
 - PPI-172 - Improve capturing and refunding process